Bootstrap 3.0 Collapse Update
Do not apply this, as it does not work yet, but I'm working to update to use the 'collapse', 'collapsing', and 'collapse in' CSS classes. Honestly, I'm guessing as to their order of use, especially the '.collapse .in' pair.

I removed the FixUpHeight function as it no longer seemed efficient to have a generic function, due to the changing CSS classes.

The major issue I'm having is that the promise resolves for expand and collapse are happening at the wrong time (e.g., when it collapses the expand .then function fires; nothing happens as isCollapsed is in the wrong state). Maybe I'm doing something dumb though. :P